src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
backoff==1.11.1
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.7.0
idna==3.10
load-dotenv==0.1.0
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
spotify==0.10.2
spotipy==2.25.1
urllib3==2.5.0
yarl==1.20.1
